+++ CHANGELOG.md
@@ -29,6 +29,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   (<https://github.com/openvinotoolkit/datumaro/pull/1055>)
 - Add CVAT data format document
   (<https://github.com/openvinotoolkit/datumaro/pull/1060>)
+- Enhance the error message for datum stats to be more user friendly
+  (<https://github.com/openvinotoolkit/datumaro/pull/1069>)
 
 ### Bug fixes
 - Fix warnings in test_visualizer.py